Revert "UAR-369: Update statement 1 (same statement as registration)"
