+++ src/config/index.ts
@@ -55,7 +55,6 @@ export const BENEFICIAL_OWNER_TYPE_PAGE = "beneficial-owner-type";
 export const UPDATE_BENEFICIAL_OWNER_TYPE_PAGE = "update-beneficial-owner-type";
 export const UPDATE_BENEFICIAL_OWNER_OTHER_PAGE = "update-beneficial-owner-other";
 export const UPDATE_BENEFICIAL_OWNER_GOV_PAGE = "update-beneficial-owner-gov";
-export const UPDATE_BENEFICIAL_OWNER_STATEMENTS_PAGE = "update-beneficial-owner-statements";
 export const BENEFICIAL_OWNER_DELETE_WARNING_PAGE = "beneficial-owner-delete-warning";
 export const CANNOT_USE_PAGE = "cannot-use";
 export const CHECK_YOUR_ANSWERS_PAGE = "check-your-answers";
@@ -188,7 +187,6 @@ export const UPDATE_INTERRUPT_CARD_URL = UPDATE_AN_OVERSEAS_ENTITY_URL + UPDATE_
 export const UPDATE_SIGN_OUT_URL = UPDATE_AN_OVERSEAS_ENTITY_URL + UPDATE_SIGN_OUT_PAGE;
 export const UPDATE_MANAGING_OFFICER_URL = UPDATE_AN_OVERSEAS_ENTITY_URL + UPDATE_MANAGING_OFFICER_PAGE;
 export const UPDATE_MANAGING_OFFICER_CORPORATE_URL = UPDATE_AN_OVERSEAS_ENTITY_URL + UPDATE_MANAGING_OFFICER_CORPORATE_PAGE;
-export const UPDATE_BENEFICIAL_OWNER_STATEMENTS_URL = UPDATE_AN_OVERSEAS_ENTITY_URL + UPDATE_BENEFICIAL_OWNER_STATEMENTS_PAGE;
 export const UPDATE_REGISTRABLE_BENEFICIAL_OWNER_URL = UPDATE_AN_OVERSEAS_ENTITY_URL + UPDATE_REGISTRABLE_BENEFICIAL_OWNER_PAGE;
 
 // URL PARAMS
