+++ src/controllers/index.ts
@@ -55,7 +55,6 @@ export * as updateBeneficialOwnerBoMoReview from "./update/update.beneficial.own
 export * as overseasEntityUpdateDetails from "./update/overseas.entity.update.details.controller";
 export * as updateBeneficialOwnerType from "./update/beneficial.owner.type.controller";
 export * as updateBeneficialOwnerGov from "./update/update.beneficial.owner.gov.controller";
-export * as updateBeneficialOwnerStatements from "./update/update.beneficial.owner.statements.controller";
 export * as updateManagingOfficerIndividual from "./update/update.managing.officer.controller";
 export * as updateBeneficialOwnerIndividual from "./update/update.beneficial.owner.individual.controller";
 export * as updateSignOut from "./update/update.sign.out.controller";
