+++ src/controllers/update/overseas.entity.review.controller.ts
@@ -34,7 +34,7 @@ export const get = (req: Request, res: Response, next: NextFunction) => {
 export const post = (req: Request, res: Response, next: NextFunction) => {
   try {
     logger.debugRequest(req, `POST ${config.OVERSEAS_ENTITY_REVIEW_PAGE}`);
-    return res.redirect(config.UPDATE_BENEFICIAL_OWNER_STATEMENTS_PAGE);
+    return res.redirect(config.UPDATE_REGISTRABLE_BENEFICIAL_OWNER_PAGE);
   } catch (errors) {
     logger.errorRequest(req, errors);
     next(errors);
