--- src/controllers/update/update.beneficial.owner.statements.controller.ts
@@ -1,11 +1,0 @@
-import { NextFunction, Request, Response } from "express";
-
-import { getBeneficialOwnerStatements, postBeneficialOwnerStatements } from "../../utils/beneficial.owner.statements";
-
-export const get = (req: Request, res: Response, next: NextFunction) => {
-  getBeneficialOwnerStatements(req, res, next, false);
-};
-
-export const post = (req: Request, res: Response, next: NextFunction) => {
-  postBeneficialOwnerStatements(req, res, next, false);
-};
