+++ src/utils/navigation.ts
@@ -73,7 +73,7 @@ export const NAVIGATION: Navigation = {
   [config.OVERSEAS_ENTITY_REVIEW_URL]: {
     currentPage: config.OVERSEAS_ENTITY_REVIEW_PAGE,
     previousPage: () => config.UPDATE_OVERSEAS_ENTITY_CONFIRM_URL,
-    nextPage: [config.UPDATE_BENEFICIAL_OWNER_STATEMENTS_URL]
+    nextPage: [config.UPDATE_REGISTRABLE_BENEFICIAL_OWNER_URL]
   },
   [config.UPDATE_BENEFICIAL_OWNER_BO_MO_REVIEW_URL]: {
     currentPage: config.UPDATE_BENEFICIAL_OWNER_BO_MO_REVIEW_PAGE,
@@ -94,11 +94,6 @@ export const NAVIGATION: Navigation = {
     currentPage: config.UPDATE_CHECK_YOUR_ANSWERS_PAGE,
     previousPage: () => config.OVERSEAS_ENTITY_REVIEW_URL,
     nextPage: []
-  },
-  [config.UPDATE_BENEFICIAL_OWNER_STATEMENTS_URL]: {
-    currentPage: config.UPDATE_BENEFICIAL_OWNER_STATEMENTS_PAGE,
-    previousPage: () => config.OVERSEAS_ENTITY_REVIEW_URL,
-    nextPage: [config.UPDATE_REGISTRABLE_BENEFICIAL_OWNER_URL]
   },
   [config.SECURE_REGISTER_FILTER_URL]: {
     currentPage: config.SECURE_REGISTER_FILTER_PAGE,
