+++ src/validation/index.ts
@@ -24,7 +24,6 @@ import { trustDetails } from "./trust.details.validation";
 import { trustIndividualBeneficialOwner } from "./trust.individual.beneficial.owner.validation";
 import { trustHistoricalBeneficialOwner } from "./trust.historical.beneficial.owner.validation";
 import { addTrustValidations } from "./add.trust.validation";
-import { updateBeneficialOwnerStatements } from "./update.beneficial.owner.statements.validation";
 import { trustLegalEntityBeneficialOwnerValidator } from "./trust-legal-entity-beneficial-owner.validation";
 import { registrableBeneficialOwner } from "./registrable.beneficial.owner.validation";
 
@@ -54,7 +53,6 @@ export const validator = {
   trustInvolved,
   trustDetails,
   updateBeneficialOwnerAndManagingOfficerType,
-  updateBeneficialOwnerStatements,
   trustIndividualBeneficialOwner,
   trustHistoricalBeneficialOwner,
   addTrust: addTrustValidations,
