+++ views/beneficial-owner-statements.html
@@ -2,6 +2,83 @@
 
 {% set title = "Has the overseas entity identified any registrable beneficial owners?" %}
 
+{% block pageTitle %}
+  {% include "includes/page-title.html" %}
+{% endblock %}
+
+{% block backLink %}
+  {% include "includes/back-link.html" %}
+{% endblock %}
+
 {% block content %}
-  {% include "includes/page-templates/beneficial-owner-statements.html" %}
+
+  <div class="govuk-grid-row">
+    <div class="govuk-grid-column-two-thirds">
+
+      {% include "includes/list/errors.html" %}
+
+      <h1 class="govuk-heading-xl">{{ title }}</h1>
+
+      <form method="post">
+
+        {{ govukRadios({
+          errorMessage: errors.beneficial_owners_statement if errors,
+          idPrefix: "beneficial_owners_statement",
+          name: "beneficial_owners_statement",
+          fieldset: {
+            legend: {
+              text: title,
+              isPageHeading: false,
+              classes: "govuk-visually-hidden"
+            }
+          },
+          items: [
+            {
+              value: "ALL_IDENTIFIED_ALL_DETAILS",
+              text: "All beneficial owners have been identified and I can provide all the required information",
+              checked: beneficial_owners_statement == "ALL_IDENTIFIED_ALL_DETAILS",
+              attributes: {
+                "data-event-id": "all-identified-all-details-radio-option"
+              }
+            },
+            {
+              value: "SOME_IDENTIFIED_ALL_DETAILS",
+              text: "Some beneficial owners have been identified and I can provide all required information",
+              hint: {
+                text: "You will also need to provide information about at least one managing officer"
+              },
+              checked: beneficial_owners_statement == "SOME_IDENTIFIED_ALL_DETAILS",
+              attributes: {
+                "data-event-id": "some-identified-all-details-radio-option"
+              }
+            },
+            {
+              value: "NONE_IDENTIFIED",
+              text: "No beneficial owners have been identified",
+              hint: {
+                text:"You will need to provide information about at least one managing officer"
+              },
+              checked: beneficial_owners_statement == "NONE_IDENTIFIED",
+              attributes: {
+                "data-event-id": "none-identified-radio-option"
+              }
+            }
+          ]
+        }) }}
+
+        {{ govukDetails({
+          summaryText: "What are registrable beneficial owners and managing officers?",
+          attributes: {
+            "data-event-id": "what-is-a-beneficial-owner-details"
+          },
+        html: "<p>A beneficial owner is any legal body that has significant influence or control over the overseas entity. It can be an individual person, other legal entity, government or public authority.</p><p>A beneficial owner is registrable if they meet one or more of the conditions known as the natures of control.</p><p>A managing officer is a director, manager or company secretary of the overseas entity.</p>"
+        }) }}
+
+        {% include "includes/save-and-continue-button.html" %}
+
+      </form>
+
+    </div>
+  </div>
+
 {% endblock %}
