--- views/update/update-beneficial-owner-statements.html
@@ -1,7 +1,0 @@
-{% extends "layout.html" %}
-
-{% set title = "Has the overseas entity identified any registrable beneficial owners?" %}
-
-{% block content %}
-  {% include "includes/page-templates/beneficial-owner-statements.html" %}
-{% endblock %}
